+++ src/pytorch_lightning/loops/epoch/evaluation_epoch_loop.py
@@ -26,7 +26,7 @@ from pytorch_lightning.utilities.auto_restart import (
     _collect_states_on_rank_zero_over_collection,
     _reload_dataloader_state_dict,
 )
-from pytorch_lightning.utilities.exceptions import ExitGracefullyException, MisconfigurationException
+from pytorch_lightning.utilities.exceptions import MisconfigurationException, SIGTERMException
 from pytorch_lightning.utilities.fetching import AbstractDataFetcher, DataLoaderIterDataFetcher
 from pytorch_lightning.utilities.imports import _fault_tolerant_training
 from pytorch_lightning.utilities.model_helpers import is_overridden
@@ -169,7 +169,7 @@ class _EvaluationEpochLoop(_Loop):
             self._outputs.append(output)
 
         if not self.batch_progress.is_last_batch and self.trainer.received_sigterm:
-            raise ExitGracefullyException
+            raise SIGTERMException
 
     def on_run_end(self) -> EPOCH_OUTPUT:
         """Returns the outputs of the whole run."""